doso-ai-package/pyproject.toml
[project]
name = "doso-ai"
version = "0.1.0"
description = "Dealer Inventory Optimization System AI"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "Ford DOSO Team", email = "dosoai@ford.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.110.0",
    "openai-agents>=0.0.14",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "pydantic>=2.10, <3",
    "pandas>=2.0.0",
    "prophet>=1.1.5",
    "scikit-learn>=1.4.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pdfplumber>=0.10.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4"
]

[project.urls]
Homepage = "https://github.com/ford/doso-ai"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/**/*.py",
    "README.md",
]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212"]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["D", "S101"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["pandas.*", "prophet.*", "pdfplumber.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
